Trait sky_sl::db::SourceDatabase

pub trait SourceDatabase: Database + HasQueryGroup<SourceDatabaseStorage> + Database {
    fn input_file(&self, key0: Utf8PathBuf) -> Arc<String>;

    fn set_input_file(&mut self, key0: Utf8PathBuf, value__: Arc<String>);

    fn set_input_file_with_durability(
        &mut self, 
        key0: Utf8PathBuf, 
        value__: Arc<String>, 
        durability__: Durability
    );

    fn line_index(&self, key0: Utf8PathBuf) -> Arc<LineIndex>;
}

Required methods

fn input_file(&self, key0: Utf8PathBuf) -> Arc<String>

fn set_input_file(&mut self, key0: Utf8PathBuf, value__: Arc<String>)

Set the value of the input_file input.

See input_file for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_input_file_with_durability(
&mut self,
key0: Utf8PathBuf,
value__: Arc<String>,
durability__: Durability
)

Set the value of the input_file input and promise that its value will never change again.

See input_file for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn line_index(&self, key0: Utf8PathBuf) -> Arc<LineIndex>