Module sky_sl::syn::cst

Structs

LineIndex

LineTextRange

SyntaxId

TextPosition

TextRange
A range in text, represented as a pair of TextSize.

TextSize
A measure of text length. Also, equivalently, an index into text.

Trivia
Stores whitespace information around a Token. Tokens own the trivia after it up to the first new line Tokens own the trivia before it starting with the first new line

Enums

Associativity

Lang

Operator

SyntaxKind

WalkEvent
WalkEvent describes tree walking process.

Type Definitions

Builder

Checkpoint

GreenNode

SyntaxElement

SyntaxNode

SyntaxNodeChildren

SyntaxToken