Enum sky_sl::syn::cst::SyntaxKind

#[repr(u16)]
pub enum SyntaxKind {

    Module,
    Struct,
    MemberList,
    Member,
    Fn,
    FnSignature,
    ArgumentList,
    Argument,
    ReturnType,
    Block,
    LetStatement,
    ExpressionStatement,
    GroupExpression,
    PathExpression,
    FieldAccessExpression,
    LiteralExpression,
    BinaryExpression,
    CallExpression,
    CallArgumentList,
    IndexExpression,
    StructExpression,
    StructExpressionFields,
    StructExpressionField,
    Operator,
    LetKeyword,
    StructKeyword,
    FnKeyword,
    Identifier,
    TypeIdentifier,
    Whitespace,
    Comment,
    NumLiteral,
    Semicolon,
    Comma,
    Dot,
    OpenParen,
    CloseParen,
    OpenBrace,
    CloseBrace,
    OpenBracket,
    CloseBracket,
    At,
    Pound,
    Tilde,
    Question,
    Colon,
    Dollar,
    Equals,
    Bang,
    LessThan,
    GreatherThan,
    Minus,
    And,
    VerticalBar,
    Plus,
    Star,
    Slash,
    Caret,
    Percent,
    EOF,
    Error,

}

Variants

Module

A module

Struct

A struct

MemberList

A struct member list

Member

A struct member

Fn

A function

FnSignature

A function signature e.g. fn example(argument: ArgumentType) -> ReturnType

ArgumentList

A list of arguments

Argument

A single argument

ReturnType

A return type declaration e.g. -> SomeType

Block

Block

LetStatement

A let binding statement e.g. let a = 1 + 2;

ExpressionStatement

A statement consisting of a single expression e.g. my_fn();

GroupExpression

A grouping expression e.g. (a + b)

PathExpression

A path expression e.g. a, a::b

FieldAccessExpression

A postfix field access expression e.g. a.b

LiteralExpression

A literal expression

BinaryExpression

A binary expression with two operands

CallExpression

A postfix call expression e.g. a() or a.b()

CallArgumentList

A list of expressions used in a call e.g. (1.0, 2.0)

IndexExpression

A postfix index expression e.g. a[0]

StructExpression

A struct expression e.g. MyStruct { a: a, b: b }

StructExpressionFields

A list of initializers for a struct expression e.g. { a: 1.0, b: 2.0 }

StructExpressionField

A single struct expression field initializer e.g. a: 10

Operator

An operator as part of an expression

LetKeyword

the “let” keyword

StructKeyword

the “struct” keyword

FnKeyword

the “fn” keyword

Identifier

any non-keyword identifier

TypeIdentifier

any type-identifier (TODO remove when we have paths)

Whitespace

any whitespace

Comment

any comment

NumLiteral

A numeric literal (TODO: split into different typed and untyped literals, e.g. 1f16, 0.0 or 42)

Semicolon

“;”

Comma

“,”

Dot

“.”

OpenParen

“(”

CloseParen

“)”

OpenBrace

“{”

CloseBrace

“}”

OpenBracket

“[”

CloseBracket

“]”

At

“@”

Pound

“#”

Tilde

“~”

Question

“?”

Colon

“:”

Dollar

“$”

Equals

“=”

Bang

“!”

LessThan

“<”

GreatherThan

“>”

Minus

“-”

And

“&”

VerticalBar

“|”

Plus

“+”

Star

“*”

Slash

“/”

Caret

“^”

Percent

“%”

EOF

End of file

Error

Error

Implementations

impl SyntaxKind

pub fn is_keyword(self) -> bool

pub fn is_puctuation(self) -> bool

pub fn is_literal(self) -> bool

pub fn is_whitespace(self) -> bool

pub fn is_comment(self) -> bool

pub fn operator(self) -> Option<Operator>

pub fn from_keyword(ident: &str) -> Option<Self>